Move dir of licenseUtil.ts, rename thirdparty_license.txt
