+++ packages/akashic-cli-export/src/__tests__/licenseUtilSpec.ts
@@ -4,7 +4,8 @@ import { writeLicenseTextFile } from "../licenseUtil.js";
 
 
 describe("licenseUtil", () => {
-	const fixturesDir = path.resolve(__dirname, "..", "..", "__tests__", "fixtures");
+	const fixturesDir = path.resolve(__dirname, "..", "__tests__", "fixtures");
+	console.log("+ fixturesDir:", fixturesDir);
 	const contentPath =  path.resolve(fixturesDir, "simple_game_using_external");
 	const destDir = path.resolve(fixturesDir, "output");	
 	const moduleFilePaths = [
@@ -26,7 +27,7 @@ describe("licenseUtil", () => {
 		expect(consoleSpy).toBeCalledWith(expect.stringMatching(/^\[WARNING\].+hoge.+LGPL-3.0-or-later.+.$/));
 		expect(result).toBeTruthy();
 
-		const license = fsx.readFileSync(path.join(destDir, "thirdparty_license.txt")).toString().split(/\r?\n/g);
+		const license = fsx.readFileSync(path.join(destDir, "library_license.txt")).toString().split(/\r?\n/g);
 		expect(license).toEqual(
 			[
 				"# external",
