+++ packages/akashic-cli-export/src/zip/__tests__/convertSpec.ts
@@ -8,7 +8,7 @@ import { vi } from "vitest";
 import { validateGameJson } from "../../utils.js";
 import type { ConvertGameParameterObject } from "../convert.js";
 import { bundleScripts, convertGame, validateGameJsonForNicolive } from "../convert.js";
-import { LICENSE_TEXT_PREFIX } from "../licenseUtil.js";
+import { LICENSE_TEXT_PREFIX } from "../../licenseUtil.js";
 
 const require = createRequire(import.meta.url);
 const fixturesDir = path.resolve(__dirname, "..", "..", "__tests__", "fixtures");
@@ -113,7 +113,7 @@ describe("convert", () => {
 					expect(fs.existsSync(path.join(destDir, "script/bar.js"))).toBe(true);
 					expect(fs.existsSync(path.join(destDir, "script/foo.js"))).toBe(true);
 					expect(fs.existsSync(path.join(destDir, "game.json"))).toBe(true);
-					expect(fs.existsSync(path.join(destDir, "thirdparty_license.txt"))).toBeFalsy();
+					expect(fs.existsSync(path.join(destDir, "library_license.txt"))).toBeFalsy();
 				});
 		});
 
@@ -132,7 +132,7 @@ describe("convert", () => {
 					expect(fs.existsSync(path.join(destDir, "text/test.json"))).toBe(true);
 					expect(fs.existsSync(path.join(destDir, "game.json"))).toBe(true);
 					expect(fs.existsSync(path.join(destDir, "package.json"))).toBe(true);
-					expect(fs.existsSync(path.join(destDir, "thirdparty_license.txt"))).toBeTruthy();
+					expect(fs.existsSync(path.join(destDir, "library_license.txt"))).toBeTruthy();
 				});
 		});
 
@@ -152,7 +152,7 @@ describe("convert", () => {
 					expect(fs.existsSync(path.join(destDir, "text/test.json"))).toBe(true);
 					expect(fs.existsSync(path.join(destDir, "game.json"))).toBe(true);
 					expect(fs.existsSync(path.join(destDir, "package.json"))).toBe(false);
-					expect(fs.existsSync(path.join(destDir, "thirdparty_license.txt"))).toBeTruthy();
+					expect(fs.existsSync(path.join(destDir, "library_license.txt"))).toBeTruthy();
 				});
 		});
 
@@ -174,7 +174,7 @@ describe("convert", () => {
 					expect(fs.existsSync(path.join(destDir, "text/test.json"))).toBe(false);
 					expect(fs.existsSync(path.join(destDir, "game.json"))).toBe(true);
 					expect(fs.existsSync(path.join(destDir, "package.json"))).toBe(true);
-					expect(fs.existsSync(path.join(destDir, "thirdparty_license.txt"))).toBeTruthy();
+					expect(fs.existsSync(path.join(destDir, "library_license.txt"))).toBeTruthy();
 					const gameJson = JSON.parse(fs.readFileSync(path.join(destDir, "game.json")).toString());
 					expect(gameJson.main).toBe("./script/aez_bundle_main.js");
 					expect(gameJson.assets.aez_bundle_main.path).toBe("script/aez_bundle_main.js");
@@ -201,7 +201,7 @@ describe("convert", () => {
 					expect(fs.existsSync(path.join(outputDirectory, "game.json"))).toBe(true);
 					expect(fs.existsSync(path.join(outputDirectory, "package.json"))).toBe(true);
 					expect(fs.existsSync(path.join(outputDirectory, "output"))).toBe(false);
-					expect(fs.existsSync(path.join(destDir, "thirdparty_license.txt"))).toBeFalsy();
+					expect(fs.existsSync(path.join(destDir, "library_license.txt"))).toBeFalsy();
 					expect(fs.readFileSync(path.join(outputDirectory, "game.json")).toString())
 						.not.toBe(fs.readFileSync(path.join(param.source, "game.json")).toString());
 				})
@@ -229,7 +229,7 @@ describe("convert", () => {
 					expect(fs.existsSync(path.join(destDir, "package.json"))).toBe(false);
 					expect(fs.readFileSync(path.join(destDir, "game.json")).toString())
 						.not.toBe(fs.readFileSync(path.join(param.source, "game.json")).toString());
-					expect(fs.existsSync(path.join(destDir, "thirdparty_license.txt"))).toBeTruthy();
+					expect(fs.existsSync(path.join(destDir, "library_license.txt"))).toBeTruthy();
 				});
 		});
 
@@ -251,7 +251,7 @@ describe("convert", () => {
 					const gameJson = fs.readFileSync(path.join(destDir, "game.json")).toString();
 					const gameJsonObj = JSON.parse(gameJson);
 					expect(gameJsonObj.assets.ignore2.global).toBeTruthy();
-					expect(fs.existsSync(path.join(destDir, "thirdparty_license.txt"))).toBeTruthy();
+					expect(fs.existsSync(path.join(destDir, "library_license.txt"))).toBeTruthy();
 				});
 		});
 
@@ -269,7 +269,7 @@ describe("convert", () => {
 					expect(fs.existsSync(path.join(destDir, "text/utf8.txt"))).toBe(true);
 					expect(fs.existsSync(path.join(destDir, "game.json"))).toBe(true);
 					expect(fs.existsSync(path.join(destDir, "package.json"))).toBe(true);
-					expect(fs.existsSync(path.join(destDir, "thirdparty_license.txt"))).toBeFalsy();
+					expect(fs.existsSync(path.join(destDir, "library_license.txt"))).toBeFalsy();
 
 					// UTF8 として読み込めることを確認
 					const main = fs.readFileSync(path.join(destDir, "script/main.js"), { encoding: "utf-8" }).toString();
@@ -312,7 +312,7 @@ describe("convert", () => {
 					expect(fs.existsSync(path.join(destDir, "text/test.json"))).toBe(false);
 					expect(fs.existsSync(path.join(destDir, "game.json"))).toBe(true);
 					expect(fs.existsSync(path.join(destDir, "package.json"))).toBe(true);
-					expect(fs.existsSync(path.join(destDir, "thirdparty_license.txt"))).toBeFalsy();
+					expect(fs.existsSync(path.join(destDir, "library_license.txt"))).toBeFalsy();
 					expect(fs.readFileSync(path.join(destDir, "game.json")).toString())
 						.not.toBe(fs.readFileSync(path.join(param.source, "game.json")).toString());
 					expect(fs.readFileSync(path.join(destDir, "script/aez_bundle_main.js")).toString())
@@ -340,7 +340,7 @@ describe("convert", () => {
 					expect(fs.existsSync(path.join(destDir, "package.json"))).toBe(true);
 					expect(fs.readFileSync(path.join(destDir, "game.json")).toString())
 						.not.toBe(fs.readFileSync(path.join(param.source, "game.json")).toString());
-					expect(fs.existsSync(path.join(destDir, "thirdparty_license.txt"))).toBeFalsy();
+					expect(fs.existsSync(path.join(destDir, "library_license.txt"))).toBeFalsy();
 				});
 		});
 
@@ -360,7 +360,7 @@ describe("convert", () => {
 					expect(fs.existsSync(path.join(destDir, "image/akashic-cli.png"))).toBe(true);
 					expect(fs.existsSync(path.join(destDir, "game.json"))).toBe(true);
 					expect(fs.existsSync(path.join(destDir, "package.json"))).toBe(true);
-					expect(fs.existsSync(path.join(destDir, "thirdparty_license.txt"))).toBeFalsy();
+					expect(fs.existsSync(path.join(destDir, "library_license.txt"))).toBeFalsy();
 					const gameJson = JSON.parse(fs.readFileSync(path.join(destDir, "game.json")).toString());
 					expect(gameJson.assets.aez_bundle_main0.path).toBe("script/aez_bundle_main0.js");
 					expect(gameJson.assets.aez_bundle_main0.type).toBe("script");
@@ -386,7 +386,7 @@ describe("convert", () => {
 					expect(fs.existsSync(path.join(destDir, "script/mainScene0.js"))).toBe(true);
 					expect(fs.existsSync(path.join(destDir, "game.json"))).toBe(true);
 					expect(fs.existsSync(path.join(destDir, "package.json"))).toBe(true);
-					expect(fs.existsSync(path.join(destDir, "thirdparty_license.txt"))).toBeFalsy();
+					expect(fs.existsSync(path.join(destDir, "library_license.txt"))).toBeFalsy();
 					expect(fs.readFileSync(path.join(destDir, "script/mainScene.js")).toString())
 						.toBe(fs.readFileSync(path.join(param.source, "script/mainScene.js")).toString());
 					const gameJson = JSON.parse(fs.readFileSync(path.join(destDir, "game.json")).toString());
@@ -412,7 +412,7 @@ describe("convert", () => {
 					expect(fs.existsSync(path.join(destDir, "text/test.json"))).toBe(false);
 					expect(fs.existsSync(path.join(destDir, "game.json"))).toBe(true);
 					expect(fs.existsSync(path.join(destDir, "package.json"))).toBe(true);
-					expect(fs.existsSync(path.join(destDir, "thirdparty_license.txt"))).toBeFalsy();
+					expect(fs.existsSync(path.join(destDir, "library_license.txt"))).toBeFalsy();
 
 					const gameJson = JSON.parse(fs.readFileSync(path.join(destDir, "game.json")).toString());
 					expect(gameJson.assets.aez_bundle_main.path).toBe("script/aez_bundle_main.js");
@@ -450,7 +450,7 @@ describe("convert", () => {
 					expect(fs.existsSync(path.join(destDir, "script/bar.js"))).toBe(true);
 					expect(fs.existsSync(path.join(destDir, "script/foo.js"))).toBe(false);
 					expect(fs.existsSync(path.join(destDir, "text/test.json"))).toBe(false);
-					expect(fs.existsSync(path.join(destDir, "thirdparty_license.txt"))).toBeFalsy(); // node_modules に LICENSE がないので false
+					expect(fs.existsSync(path.join(destDir, "library_license.txt"))).toBeFalsy(); // node_modules に LICENSE がないので false
 
 					const gameJson = JSON.parse(fs.readFileSync(path.join(destDir, "game.json")).toString());
 					expect(gameJson.assets.aez_bundle_main.path).toBe("script/aez_bundle_main.js");
@@ -487,7 +487,7 @@ describe("convert", () => {
 					expect(fs.existsSync(path.join(destDir, "image/akashic-cli.png"))).toBe(true);
 					expect(fs.existsSync(path.join(destDir, "game.json"))).toBe(true);
 					expect(fs.existsSync(path.join(destDir, "package.json"))).toBe(true);
-					expect(fs.existsSync(path.join(destDir, "thirdparty_license.txt"))).toBeFalsy();
+					expect(fs.existsSync(path.join(destDir, "library_license.txt"))).toBeFalsy();
 					const gameJson = JSON.parse(fs.readFileSync(path.join(destDir, "game.json")).toString());
 					const imgAsset = gameJson.assets.aez_bundle_main;
 					expect(imgAsset.type).toBe("image");
@@ -521,7 +521,7 @@ describe("convert", () => {
 			return convertGame(param)
 				.then(() => {
 					expect(fs.existsSync(path.join(destDir, "script/aez_bundle_main.js"))).toBe(true);
-					expect(fs.existsSync(path.join(destDir, "thirdparty_license.txt"))).toBeFalsy(); // node_modules に LICENSE がないので false
+					expect(fs.existsSync(path.join(destDir, "library_license.txt"))).toBeFalsy(); // node_modules に LICENSE がないので false
 					// bundle済みのファイルは残らない
 					expect(fs.existsSync(path.join(destDir, "script/main.js"))).toBe(false);
 					expect(fs.existsSync(path.join(destDir, "script/foo.js"))).toBe(false);
@@ -553,7 +553,7 @@ describe("convert", () => {
 					expect(fs.existsSync(path.join(destDir, "script/aez_asset_bundle.js"))).toBe(true);
 					expect(fs.existsSync(path.join(destDir, "game.json"))).toBe(true);
 					expect(fs.existsSync(path.join(destDir, "package.json"))).toBe(true);
-					expect(fs.existsSync(path.join(destDir, "thirdparty_license.txt"))).toBeFalsy();
+					expect(fs.existsSync(path.join(destDir, "library_license.txt"))).toBeFalsy();
 					const gameJson = JSON.parse(fs.readFileSync(path.join(destDir, "game.json")).toString());
 					expect(gameJson.assets.aez_asset_bundle.path).toBe("script/aez_asset_bundle.js");
 					expect(gameJson.assets.aez_asset_bundle.type).toBe("script");
@@ -616,7 +616,7 @@ describe("convert", () => {
 				});
 		});
 
-		it("thirdparty_license.txt by bundle", () => {
+		it("library_license.txt by bundle", () => {
 			const param = {
 				source: path.resolve(fixturesDir, "simple_game_using_external"),
 				dest: destDir,
@@ -627,14 +627,14 @@ describe("convert", () => {
 					expect(fs.existsSync(path.join(destDir, "node_modules/external/index.js"))).toBeFalsy();
 					expect(fs.existsSync(path.join(destDir, "node_modules/external/package.json"))).toBeTruthy();
 					expect(fs.existsSync(path.join(destDir, "script/main.js"))).toBeFalsy();
-					expect(fs.existsSync(path.join(destDir, "thirdparty_license.txt"))).toBeTruthy();
+					expect(fs.existsSync(path.join(destDir, "library_license.txt"))).toBeTruthy();
 
 					const mainScript = fs.readFileSync(path.join(destDir, "script/aez_bundle_main.js")).toString().split("\n");
 					expect(mainScript[0]).toBe(LICENSE_TEXT_PREFIX.replace(/\r?\n/g, ""));
 				});
 		});
 
-		it("thirdparty_license.txt by no bundle", () => {
+		it("library_license.txt by no bundle", () => {
 			const param = {
 				source: path.resolve(fixturesDir, "simple_game_using_external"),
 				dest: destDir,
@@ -644,7 +644,7 @@ describe("convert", () => {
 					expect(fs.existsSync(path.join(destDir, "node_modules/external/index.js"))).toBeTruthy();
 					expect(fs.existsSync(path.join(destDir, "node_modules/external/package.json"))).toBeTruthy();
 					expect(fs.existsSync(path.join(destDir, "script/main.js"))).toBeTruthy();
-					expect(fs.existsSync(path.join(destDir, "thirdparty_license.txt"))).toBeTruthy();
+					expect(fs.existsSync(path.join(destDir, "library_license.txt"))).toBeTruthy();
 
 					const mainScript = fs.readFileSync(path.join(destDir, "script/main.js")).toString().split("\n");
 					expect(mainScript[0]).toBe(LICENSE_TEXT_PREFIX.replace(/\r?\n/g, ""));
