+++ packages/akashic-cli-export/src/zip/convert.ts
@@ -16,7 +16,7 @@ import { validateGameJson } from "../utils.js";
 import { getFromHttps } from "./apiUtil.js";
 import { NICOLIVE_SIZE_LIMIT_GAME_JSON, NICOLIVE_SIZE_LIMIT_TOTAL_FILE } from "./constants.js";
 import * as gcu from "./GameConfigurationUtil.js";
-import * as liceneUtil from "./licenseUtil.js";
+import * as liceneUtil from "../licenseUtil.js";
 import { transformPackSmallImages } from "./transformPackImages.js";
 
 // NOTE: 以下のパッケージは型定義が存在しないか JS と型定義の齟齬があるため `require()` を用いている
